requirements.txt
asgiref==3.2.7
Django==3.0.5
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-password-reset==2.0
django-taggit==1.2.0
Markdown==3.2.1
Pillow==7.1.1
Pygments==2.6.1
pytz==2019.3
sqlparse==0.3.1
